+++ src/main/java/space/space_spring/domain/authorization/jwt/model/JwtLoginProvider.java
@@ -1,10 +1,9 @@
-package space.space_spring.jwt;
+package space.space_spring.domain.authorization.jwt.model;
 
 import io.jsonwebtoken.*;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import space.space_spring.domain.authorization.jwt.model.TokenType;
 import space.space_spring.entity.User;
 import space.space_spring.exception.CustomException;
 import space.space_spring.exception.jwt.bad_request.JwtUnsupportedTokenException;
