+++ src/main/java/space/space_spring/domain/authorization/jwt/repository/JwtRepository.java
@@ -1,7 +1,11 @@
 package space.space_spring.domain.authorization.jwt.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 import space.space_spring.entity.TokenStorage;
 import space.space_spring.entity.User;
 
@@ -11,5 +15,9 @@ import java.util.Optional;
 public interface JwtRepository extends JpaRepository<TokenStorage, Long> {
 
     Optional<TokenStorage> findByUser(User user);
-    void deleteByUser(User user);
+
+    @Transactional
+    @Modifying
+    @Query("DELETE FROM TokenStorage t WHERE t.user = :user")
+    void deleteByUser(@Param("user") User user);
 }
